# backend/internal/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/audit": {
            "get": {
                "description": "Returns the audit trail entries.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Audit"
                ],
                "summary": "List Audit Entries",
                "responses": {
                    "200": {
                        "description": "Audit entry list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/dags": {
            "get": {
                "description": "Returns all DAG configurations.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAG"
                ],
                "summary": "List DAGs",
                "responses": {
                    "200": {
                        "description": "DAG list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new DAG configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAG"
                ],
                "summary": "Create DAG",
                "parameters": [
                    {
                        "description": "DAG configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DAGConfig"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created DAG",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/dags/{id}": {
            "get": {
                "description": "Returns a DAG configuration by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAG"
                ],
                "summary": "Get DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DAG",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing DAG configuration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAG"
                ],
                "summary": "Update DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "DAG configuration",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DAGConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated DAG",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a DAG configuration by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DAG"
                ],
                "summary": "Delete DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.DeleteDAGResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
        "/api/v1/dags/{id}/run": {
            "post": {
                "description": "Starts execution for a saved DAG.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Execution"
                ],
                "summary": "Run DAG",
                "parameters": [
                    {
                        "type": "string",
                        "description": "DAG ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/dto.RunExecutionResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/experiences": {
            "get": {
                "description": "Returns stored memory experiences.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Memory"
                ],
                "summary": "List Experiences",
                "responses": {
                    "200": {
                        "description": "Experience list",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "dto.DeleteDAGResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.RunExecutionResponse": {
            "type": "object",
            "properties": {
                "execution_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ExecutionStatus"
                }
            }
        },
...
        "models.DAGConfig": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Edge"
                    }
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Node"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Edge": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "JSON path expression for conditional routing",
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "models.ExecutionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "timeout",
                "suspended"
            ],
            "x-enum-comments": {
                "StatusSuspended": "waiting for human approval"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "waiting for human approval"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusRunning",
                "StatusCompleted",
                "StatusFailed",
                "StatusTimeout",
                "StatusSuspended"
            ]
        },
...
        "models.Node": {
            "type": "object",
            "properties": {
                "action": {
                    "description": "bash command, LLM prompt template, MCP tool name, etc.",
                    "type": "string"
                },
                "config": {
                    "description": "additional per-type configuration",
                    "type": "object",
                    "additionalProperties": true
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.NodeType"
                }
            }
        },
        "models.NodeType": {
            "type": "string",
            "enum": [
                "script",
                "llm",
                "mcp",
                "human",
                "router",
                "web_interaction"
            ],
            "x-enum-varnames": [
                "NodeTypeScript",
                "NodeTypeLLM",
                "NodeTypeMCP",
                "NodeTypeHuman",
                "NodeTypeRouter",
                "NodeTypeWebInteraction"
            ]
        }
    }
}